+++ src/components/RichTextEditor/Plugins/ListPlugin/OrderedListPlugin/index.ts
@@ -16,3 +16,6 @@ export class OrderedListPlugin extends ListPlugin {
         });
     }
 }
+
+export * from './OrderedListButton';
+export * from './OrderedListMarkupElement';
