+++ src/components/RichTextEditor/Plugins/ListPlugin/UnorderedListPlugin/index.ts
@@ -16,3 +16,6 @@ export class UnorderedListPlugin extends ListPlugin {
         });
     }
 }
+
+export * from './UnorderedListButton';
+export * from './UnorderedListMarkupElement';
