+++ src/components/RichTextEditor/Plugins/UnderlinePlugin/index.ts
@@ -19,3 +19,5 @@ export class UnderlinePlugin extends Plugin {
         return [createUnderlinePlugin()];
     }
 }
+
+export * from './UnderlineMarkupElement';
