+++ src/components/RichTextEditor/serializer/utils/serializeNodeToHtmlRecursive.spec.ts
@@ -1,7 +1,7 @@
-import { TextStyles } from '@components/RichTextEditor/Plugins/TextStylePlugin/TextStyles';
-import { OL_CLASSES, UL_CLASSES } from '@components/RichTextEditor/components';
-import { defaultDesignTokens } from '@components/RichTextEditor/utils/defaultDesignTokens';
 import { ELEMENT_LI, ELEMENT_LINK, ELEMENT_OL, ELEMENT_PARAGRAPH, ELEMENT_UL } from '@udecode/plate';
+import { OL_CLASSES, UL_CLASSES } from '@components/RichTextEditor/Plugins';
+import { TextStyles } from '@components/RichTextEditor/Plugins/TextStylePlugin/TextStyles';
+import { defaultDesignTokens } from '@components/RichTextEditor/utils/defaultDesignTokens';
 import { serializeNodeToHtmlRecursive } from './serializeNodeToHtmlRecursive';
 
 describe('serializeNodeToHtmlRecursive()', () => {
