+++ src/components/RichTextEditor/serializer/utils/serializeNodeToHtmlRecursive.ts
@@ -1,8 +1,9 @@
-import { OL_CLASSES, UL_CLASSES } from '@components/RichTextEditor/components';
-import { ELEMENT_CHECK_ITEM } from '../../Plugins/CheckboxListPlugin/id';
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { ELEMENT_LI, ELEMENT_LINK, ELEMENT_OL, ELEMENT_PARAGRAPH, ELEMENT_UL, TDescendant } from '@udecode/plate';
+import { ELEMENT_CHECK_ITEM, OL_CLASSES, UL_CLASSES } from '@components/RichTextEditor/Plugins';
 import { DesignTokens } from '@components/RichTextEditor/types';
 import { TextStyles } from '@components/RichTextEditor/Plugins/TextStylePlugin/TextStyles';
-import { ELEMENT_LI, ELEMENT_LINK, ELEMENT_OL, ELEMENT_PARAGRAPH, ELEMENT_UL, TDescendant } from '@udecode/plate';
 import escapeHtml from 'escape-html';
 import { reactCssPropsToCss } from './reactCssPropsToCss';
 import { serializeLeafToHtml } from './serializeLeafToHtml';
